feat(ui): implemented logout from other sessions in user profile
