+++ apps/knowii/src/Components/ActionSection.tsx
@@ -1,4 +1,4 @@
-import React, { PropsWithChildren } from 'react';
+import { PropsWithChildren } from 'react';
 import SectionTitle from '@/Components/SectionTitle';
 
 interface Props {
