+++ apps/knowii/src/Components/SectionTitle.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 interface Props {
   title: string;
   description: string;
